requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
langchain>=0.0.350
langchain-groq>=0.0.1
langgraph>=0.0.60
pydantic>=2.5.0
pydantic-settings>=2.1.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0